refactor(order-list): socket connection is moved to pages; renamed endpoint prop; removed withToken prop; added isOrdersReverse prop
